+++ NEWS.md
@@ -1,6 +1,9 @@
 # gwasRtools 0.1.1
 
 * Fix internal -log10p calculation
+* Swap most "Depends" to "Imports"
+* Add dependency for {ieugwasr}
+* Add `ld_clump` option to get independent SNPs at the same locus
 
 # gwasRtools 0.1.0